+++ styles/index.scss
@@ -1,6 +1,6 @@
 main {
 
-  header {
+  header.search-wrapper {
     display: flex;
     flex-direction: column;
     justify-content: center;
