+++ styles/map-show.scss
@@ -10,7 +10,7 @@ div.pinned-map {
     div {
       display: flex;
       justify-content: space-between;
-      span.date {
+      span {
         font-size: 0.5em;
         margin-right: 0.2em;
       }
