+++ views/index.ejs
@@ -27,7 +27,7 @@
   <body>
     <%- include("partials/_header") %>
     <main>
-      <header>
+      <header class="search-wrapper">
         <form action="/maps/json">
           <label for="search">Search for a map:</label>
           <input name="search" type="text" placeholder="Enter a key word or name..." />
